% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{applyCond}
\alias{applyCond}
\title{Apply evaluation condition to result}
\usage{
applyCond(x, condition, data)
}
\arguments{
\item{x}{result to be adjusted (emptied when condition criteria are not met)}

\item{condition}{text expression in R syntax}

\item{data}{data frame to be referred, columns can be referred directly by
column name}
}
\description{
Function evaluates condition and if condition is met (TRUE),
  then result value returned. If condition is not met (FALSE) empty value
  will be returned. Condition may refer to \code{data} provided as a
  parameter
}
