% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generetor.R
\name{castValue}
\alias{castValue}
\title{Convert value into specific data-type}
\usage{
castValue(value, type)
}
\arguments{
\item{value}{value to be converted}

\item{type}{data type to be returned}
}
\description{
Convert value into specific data-type
}
