% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_sets}
\alias{demo_sets}
\title{Lists of Values for data generation}
\format{A dataframe with following variables
\describe{
  \item{Set}{
    Name of the set as it is reffered by \code{List.Of.Values} in rules for
    attribues of \code{Value.Type} = \code{LOV}
  }
  \item{Value}{
    Possible values that belong to set
  }
}}
\usage{
demo_sets
}
\description{
A dataset contains lists of values that are referred by \link{demo_rules}
}
\keyword{datasets}
