% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generateAttr}
\alias{generateAttr}
\title{Generate values}
\usage{
generateAttr(count, attr_type, attr_len = NULL, attr_num_dec = NULL,
  eval_cond = NULL, value_type = c("Empty", "Fixed", "LOV", "Random",
  "Expression"), fix_offset_value = NULL, lov = NULL,
  rand_dist_name = NULL, rand_dist_mean = NULL, rand_dist_sd = NULL,
  sign_type = NULL, expression = NULL, data = NULL, seed = NULL)
}
\arguments{
\item{count}{number of values to generate}

\item{attr_type}{data type of attribute (value).}

\item{attr_len}{maximum length of attribute}

\item{attr_num_dec}{maximum number of digits in decimal part of number}

\item{eval_cond}{evaluation condition (R syntax, can refer to other columns
in \code{data} param)}

\item{value_type}{mode to generate value}

\item{fix_offset_value}{fixed value to be returned / offset for random
numbers or dates}

\item{lov}{vector of exact values that can be used for value_type = "LOV"}

\item{rand_dist_name}{distribution name for random number generation}

\item{rand_dist_mean}{distribution mean for random number generation}

\item{rand_dist_sd}{distribution standard deviation for random number
generation}

\item{sign_type}{type of sign for randomly generated number (see
\link{validateSign})}

\item{expression}{evaluation expression (R syntax, can refer to other columns
in \code{data} param)}

\item{data}{already generated result in form of data frame, so it can be
referred by \code{eval_cond} or \code{expression}}

\item{seed}{integer seed for random number generator; using NULL seed causes
non-reproducible random generation (see \code{set.seed()} help)}
}
\description{
Generate values
}
