% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSetsFromDF}
\alias{makeSetsFromDF}
\title{Vector of set names and vector of set values into list of set values}
\usage{
makeSetsFromDF(set_names, set_values)
}
\arguments{
\item{set_names}{vector of set names}

\item{set_values}{vector of set values}
}
\description{
Function transforms dataframe columns (Set name and Set value)
  into list of set values. Sets may be accessed by list element names. The
  output of function can be used as and input for data generation functions
  (like \link{processTable} or \link{processAttr}).
}
