% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nest_df}
\alias{nest_df}
\title{Break data frasme into list of data-frasmes with data grouped by provided
column}
\usage{
nest_df(df, column)
}
\arguments{
\item{df}{data frame}

\item{column}{column name within data frame that should be used for groupping}
}
\description{
Break data frasme into list of data-frasmes with data grouped by provided
column
}
