% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processAttr}
\alias{processAttr}
\title{Process rules for attribute generation}
\usage{
processAttr(rules_for_attr, lovs, count, data)
}
\arguments{
\item{rules_for_attr}{dataframe with rules for attributes}

\item{lovs}{lists of values}

\item{count}{number of records to generate}

\item{data}{data frame with earlier prepared results is useful for evaluation
conditions and evaluation expressions}
}
\description{
Process rules for attribute generation
}
