% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processRules}
\alias{processRules}
\title{Process table with rules}
\usage{
processRules(rules_df, lovs_df, count)
}
\arguments{
\item{rules_df}{data frame with rules for tables and attributes generation}

\item{lovs_df}{data frame with lists of values (lov name and lov value pairs)}

\item{count}{number of records to be generated}
}
\description{
Process table with rules
}
