% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processTable}
\alias{processTable}
\title{Process rules for specific table}
\usage{
processTable(rules_for_table, lovs, count)
}
\arguments{
\item{rules_for_table}{data frame with rules to generate attributes for one
table}

\item{lovs}{list of vectors with possible values for Value Type = "LOV" list
names should be the same as they are  mentioned in rules "Lit.Of.Values"
attribute}

\item{count}{number of records to be generated}
}
\description{
Process rules for specific table
}
