% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{reduceCount}
\alias{reduceCount}
\title{Reduce number of generated elements down to number of requested elements}
\usage{
reduceCount(x, count)
}
\arguments{
\item{x}{vector of elements}

\item{count}{requested number of elements}
}
\description{
Reduce number of generated elements down to number of requested elements
}
