% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generetor.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate result of expression}
\usage{
evaluate(expression, data)
}
\arguments{
\item{expression}{text expression in R syntax}

\item{data}{data frame to be referred, columns can be referred directly by
column name}
}
\description{
Function evaluates expression, expression may refer to
  \code{data} provided as a second parameter
}
